Updated ClientBlockPlacementPacket.kt
