+++ src/main/kotlin/packets/PacketHandler.kt
@@ -484,47 +484,6 @@ class PacketHandler(
         }
     }
 
-    @PacketReceiver
-    fun onArmSwing(packet: ClientAnimationPacket) {
-
-    }
-
-    /**
-     * Called when a client places a block
-     */
-    @PacketReceiver
-    fun onBlockPlacement(packet: ClientBlockPlacementPacket) {
-        val heldStack = client.player.getHeldItem()
-        if(heldStack.isAir || heldStack.count <= 0) return
-
-        val heldItem = client.player.getHeldItemID()
-
-        val event = BlockPlaceEvent(
-            client.player,
-            packet.hand,
-            packet.blockPos.copy(),
-            packet.face,
-            packet.cursorPositionX,
-            packet.cursorPositionY,
-            packet.cursorPositionZ,
-            packet.insideBlock
-        )
-        EventManager.fire(event)
-        if (event.isCancelled) return
-
-        when (event.face) {
-            0 -> event.blockPos.y -= 1
-            1 -> event.blockPos.y += 1
-            2 -> event.blockPos.z -= 1
-            3 -> event.blockPos.z += 1
-            4 -> event.blockPos.x -= 1
-            5 -> event.blockPos.x += 1
-        }
-
-        val block = Block.getBlockFromID(heldItem) ?: Item.getItemFromID(heldItem) ?: Block.AIR
-        handlePlacement(block, event, packet.blockPos)
-    }
-
     /**
      * Every 20 ticks the client will send an empty movement packet telling the server if the
      * client is on the ground or not
@@ -1348,247 +1307,6 @@ class PacketHandler(
         return true
     }
 
-    private fun handlePlacement(block: Any, event: BlockPlaceEvent, blockPos: BlockPositionType.BlockPosition) {
-        if(block is Item && block == Item.AIR) return
-        if(block is Block && block == Block.AIR) return
-
-        try {
-            val dir = getCardinalDirection(client.player.location.yaw)
-            val properties = modifyBlockProperties(block, dir, event)
-
-            val stateID = getStateID(block, properties)
-            if (stateID == -1) return
-
-            val clickedStateID = world.modifiedBlocks[blockPos]?.stateID ?: 0
-            val clickedItemID = clickedStateID.let { Item.getIDFromState(it) }
-
-            for (bed in BlockTags.BEDS) {
-                if (clickedItemID == bed.id) {
-                    handleBedClick(blockPos)
-                    break
-                }
-            }
-
-            players.forEach {
-                val entry = handlePlayerSpecificBlockPlacement(event, stateID, it, block)
-                world.modifiedBlocks[event.blockPos] = entry
-            }
-
-            if(client.player.gameMode == GameMode.SURVIVAL) {
-                handleItemDecrease()
-            }
-        } catch (e: IllegalArgumentException) {
-            //do nothing
-        }
-    }
-
-    private fun getStateID(block: Any, properties: Map<String, String>): Int {
-        return when (block) {
-            is Block -> Block.getStateID(block, properties)
-            is Item -> Item.getStateID(block, properties)
-            else -> -1
-        }
-    }
-
-    private fun modifyBlockProperties(
-        block: Any,
-        cardinalDirection: String,
-        event: BlockPlaceEvent
-    ): MutableMap<String, String> {
-        val properties = mutableMapOf<String, String>()
-
-        for (furnace in BlockTags.FURNANCES) {
-            if (block == furnace) {
-                properties["facing"] = cardinalDirection
-                properties["lit"] = "false"
-
-                if (block == Block.CAMPFIRE) {
-                    properties["waterlogged"] = "false"
-                    properties["signal_fire"] = "false"
-                }
-            }
-        }
-
-        if (block == Block.END_ROD) {
-            properties["facing"] = cardinalDirection
-        }
-
-        if (block == Block.GRINDSTONE) {
-            properties["facing"] = cardinalDirection
-            properties["face"] = "floor"
-        }
-
-        if (block == BlockTags.BANNERS || block == BlockTags.SKULLS) {
-            properties["rotation"] = getRotationalDirection(client.player.location.yaw).toString()
-        }
-
-        if (block == BlockTags.SIGNS) {
-            properties["rotation"] = getRotationalDirection(client.player.location.yaw).toString()
-            properties["waterlogged"] = "false"
-        }
-
-        modifyStairProperties(block, cardinalDirection).forEach { (key, value) ->
-            properties[key] = value
-        }
-
-        modifyRedstoneBlockProperties(block, cardinalDirection).forEach { (key, value) ->
-            properties[key] = value
-        }
-
-        modifyAxisAlignedBlocks(block).forEach { (key, value) ->
-            properties[key] = value
-        }
-
-        if (block is Item) {
-            modifyBedBlocks(block, cardinalDirection, event).forEach { (key, value) ->
-                properties[key] = value
-            }
-        }
-
-        return properties
-    }
-
-    private fun modifyStairProperties(block: Any, cardinalDirection: String): MutableMap<String, String> {
-        val properties = mutableMapOf<String, String>()
-
-        for (stair in BlockTags.STAIRS) {
-            if (block == stair) {
-                properties["facing"] = cardinalDirection
-                properties["half"] = "bottom"
-                properties["shape"] = "straight"
-                properties["waterlogged"] = "false"
-            }
-        }
-
-        return properties
-    }
-
-    private fun modifyRedstoneBlockProperties(block: Any, cardinalDirection: String): MutableMap<String, String> {
-        val properties = mutableMapOf<String, String>()
-
-        if (block == Block.DISPENSER || block == Block.DROPPER) {
-            properties["facing"] = cardinalDirection
-            properties["triggered"] = "false"
-        }
-
-        if (block == Block.PISTON) {
-            properties["extended"] = "false"
-            properties["facing"] = cardinalDirection
-        }
-
-        if (block == Block.OBSERVER) {
-            properties["facing"] = cardinalDirection
-            properties["powered"] = "false"
-        }
-
-        if (block == Block.REPEATER) {
-            properties["facing"] = cardinalDirection
-            properties["delay"] = "1"
-            properties["locked"] = "false"
-            properties["powered"] = "false"
-        }
-
-        if (block == Block.COMPARATOR) {
-            properties["facing"] = cardinalDirection
-            properties["mode"] = "compare"
-            properties["powered"] = "false"
-        }
-
-        if (block == Block.BARREL) {
-            properties["facing"] = cardinalDirection
-            properties["open"] = "false"
-        }
-
-        if (block == Block.LECTERN) {
-            properties["facing"] = cardinalDirection
-            properties["has_book"] = "false"
-            properties["powered"] = "false"
-        }
-
-        return properties
-    }
-
-    private fun modifyAxisAlignedBlocks(block: Any): MutableMap<String, String> {
-        val properties = mutableMapOf<String, String>()
-
-        for (log in BlockTags.LOGS) {
-            if (block == log) {
-                properties["axis"] = getAxisDirection(
-                    client.player.location.yaw,
-                    client.player.location.pitch
-                ).name.lowercase()
-            }
-        }
-
-        if (block == Item.QUARTZ_PILLAR) {
-            properties["axis"] = getAxisDirection(
-                client.player.location.yaw,
-                client.player.location.pitch
-            ).name.lowercase()
-        }
-
-        if (block == Item.CHAIN) {
-            properties["axis"] = getAxisDirection(
-                client.player.location.yaw,
-                client.player.location.pitch
-            ).name.lowercase()
-        }
-
-        if (block == Item.BONE_BLOCK) {
-            properties["axis"] = getAxisDirection(
-                client.player.location.yaw,
-                client.player.location.pitch
-            ).name.lowercase()
-        }
-
-        if (block == Item.BASALT || block == Item.POLISHED_BASALT) {
-            properties["axis"] = getAxisDirection(
-                client.player.location.yaw,
-                client.player.location.pitch
-            ).name.lowercase()
-        }
-
-        return properties
-    }
-
-    private fun modifyBedBlocks(
-        block: Item,
-        cardinalDirection: String,
-        event: BlockPlaceEvent
-    ): MutableMap<String, String> {
-        val properties = mutableMapOf<String, String>()
-
-        for (bed in BlockTags.BEDS) {
-            if (block == bed) {
-                properties["facing"] = cardinalDirection
-                properties["part"] = "foot"
-                properties["occupied"] = "false"
-
-                val headPos = when (cardinalDirection) {
-                    "north" -> event.blockPos.copy(z = event.blockPos.z - 1)
-                    "south" -> event.blockPos.copy(z = event.blockPos.z + 1)
-                    "west" -> event.blockPos.copy(x = event.blockPos.x - 1)
-                    "east" -> event.blockPos.copy(x = event.blockPos.x + 1)
-                    else -> event.blockPos
-                }
-
-                val props: MutableMap<String, String> = mutableMapOf()
-                props["facing"] = cardinalDirection
-                props["part"] = "head"
-                props["occupied"] = "false"
-
-                val stateID = Item.getStateID(block, props)
-                world.modifiedBlocks[headPos] = BlockWithMetadata(block.id, stateID)
-
-                for (player in Bullet.players) {
-                    player.sendPacket(ServerBlockChangePacket(headPos, stateID))
-                }
-            }
-        }
-
-        return properties
-    }
-
     private fun sendEntities() {
         for (entity in world.entities) {
             client.sendPacket(
@@ -1620,50 +1338,6 @@ class PacketHandler(
         }
     }
 
-    private fun handleSpawnEgg(block: Item, event: BlockPlaceEvent) {
-        val itemName = block.name.removeSuffix("_SPAWN_EGG")
-        val entity = LivingEntities.entries.find { it.name.equals(itemName, true) }
-
-        if (entity != null) {
-            val location = LocationType.Location(
-                event.blockPos.x + 0.5,
-                event.blockPos.y,
-                event.blockPos.z + 0.5
-            )
-
-            val newEntity = LivingEntity()
-            val entityType = entity.id
-
-            client.sendPacket(
-                ServerSpawnLivingEntityPacket(
-                    newEntity.entityID,
-                    newEntity.uuid,
-                    entityType,
-                    location,
-                    90f,
-                    0,
-                    0,
-                    0
-                )
-            )
-
-            world.livingEntities.add(
-                Pair(
-                    newEntity, EntityData(
-                        newEntity.uuid,
-                        location,
-                        entityType,
-                        20,
-                        90f,
-                        0,
-                        0,
-                        0,
-                    )
-                )
-            )
-        }
-    }
-
     private fun sendBlockBreakParticles(player: Player, block: Int, blockPos: BlockPositionType.BlockPosition) {
         player.sendPacket(
             ServerParticlePacket(
@@ -1677,110 +1351,6 @@ class PacketHandler(
                 25
             )
         )
-    }
-
-    private fun getCardinalDirection(yaw: Float): String {
-        val rot = (yaw % 360 + 360) % 360
-        return when {
-            rot >= 45 && rot < 135 -> "west"
-            rot >= 135 && rot < 225 -> "north"
-            rot >= 225 && rot < 315 -> "east"
-            else -> "south"
-        }
-    }
-
-    private fun getRotationalDirection(yaw: Float): Int {
-        val normalizedYaw = (yaw % 360 + 360) % 360
-        return ((normalizedYaw / 22.5).toInt() % 16)
-    }
-
-    private fun getAxisDirection(yaw: Float, pitch: Float): Axis {
-        return when {
-            pitch > 60f -> Axis.Y
-            pitch < -60f -> Axis.Y
-            abs(yaw) % 180 < 45 || abs(yaw) % 180 > 135 -> Axis.Z
-            else -> Axis.X
-        }
-    }
-
-    private fun handleBedClick(blockPos: BlockPositionType.BlockPosition) {
-        val time = client.player.world!!.timeOfDay
-        if (world.weather == 0) {
-            if (time !in 12542..23459) {
-                client.player.sendMessage(
-                    Component.text("You can only sleep at night")
-                        .color(NamedTextColor.RED)
-                )
-                return
-            }
-        } else {
-            if (time !in 12010..23991) {
-                client.player.sendMessage(
-                    Component.text("You can only sleep at night")
-                        .color(NamedTextColor.RED)
-                )
-                return
-            }
-        }
-
-        val metadata = listOf(
-            MetadataType.MetadataEntry(6.toByte(), 18, 2),
-            MetadataType.MetadataEntry(13.toByte(), 10, true to blockPos),
-        )
-
-        for (player in players) {
-            player.sendPacket(ServerEntityMetadataPacket(client.player.entityID, metadata))
-        }
-
-        client.player.world!!.sleepingPlayers += 1
-        if (canSleepNow()) handleSleeping()
-    }
-
-    private fun handleWakeUp(player: Player) {
-        if (!canSleepNow()) return
-
-        val metadata = listOf(
-            MetadataType.MetadataEntry(13.toByte(), 10, false to null),
-            MetadataType.MetadataEntry(6.toByte(), 18, 0)
-        )
-
-        for (plr in players) {
-            plr.sendPacket(ServerEntityMetadataPacket(player.entityID, metadata))
-        }
-
-        player.world!!.sleepingPlayers -= 1
-    }
-
-    private fun handleSleeping() {
-        Bullet.scope.launch {
-            delay(5.seconds)
-            val world = client.player.world!!
-            if (!canSleepNow()) return@launch
-
-            for (player in players) {
-                player.sendPacket(ServerChangeGameStatePacket(1, 0f))
-                world.weather = 0
-                player.setTimeOfDay(0)
-                handleWakeUp(player)
-            }
-        }
-    }
-
-    private fun canSleepNow(): Boolean {
-        val world = client.player.world!!
-        val time = world.timeOfDay
-        val totalPlayers = players.size
-        val sleepingPlayers = world.sleepingPlayers
-
-        return if (totalPlayers > 0 && sleepingPlayers >= totalPlayers / 2) {
-            if (world.weather == 0) {
-                time in 12542..23459
-            } else {
-                time in 12010..23991
-            }
-        } else {
-            false
-        }
     }
 
     private fun checkOrbs() {
@@ -2128,43 +1698,5 @@ class PacketHandler(
         stopBlockBreak(event.blockPos)
         sendBlockBreakParticles(client.player, stateID, event.blockPos)
         removeBlock(event.blockPos)
-    }
-
-    private fun handlePlayerSpecificBlockPlacement(
-        event: BlockPlaceEvent, stateID: Int, player: Player, block: Any
-    ): BlockWithMetadata {
-        player.sendPacket(ServerBlockChangePacket(event.blockPos, stateID))
-
-        return when {
-            block is Item && block in BlockTags.SIGNS -> {
-                client.sendPacket(ServerOpenSignEditorPacket(event.blockPos))
-                BlockWithMetadata(block.id, stateID, listOf("", "", "", ""))
-            }
-
-            block is Item && block in BlockTags.SPAWN_EGGS -> {
-                handleSpawnEgg(block, event)
-                BlockWithMetadata(block.id, stateID)
-            }
-
-            block is Block -> {
-                BlockWithMetadata(block.id, stateID)
-            }
-
-            else -> BlockWithMetadata(0, stateID)
-        }
-    }
-
-    private fun handleItemDecrease() {
-        val slot = client.player.selectedSlot
-        val inv = client.player.inventory
-        val held = inv.heldStack(slot)
-
-        if(held.count > 1) {
-            held.count--
-            client.sendPacket(ServerSetSlotPacket(0, slot + 36, held.toSlotData()))
-        } else {
-            inv.setHeldSlot(slot, null)
-            client.sendPacket(ServerSetSlotPacket(0, slot + 36, Slot.SlotData(false)))
-        }
     }
 }
